Moved auth to middleware, basic data validation
