+++ server/src/controllers/usersController.ts
@@ -3,13 +3,18 @@ import { getRepository } from 'typeorm';
 import * as bcrypt from 'bcrypt';
 import * as _ from 'lodash';
 import { User } from '../entities/User';
+import { authenticate } from '../middleware/auth';
 const router = Router();
 
-router.get('/', async (req: Request, res: Response) => {
+router.get('/', authenticate(), async (req: Request, res: Response) => {
     res.status(200).send('Sending all users');
 });
 
 router.post('/', async (req: Request, res: Response) => {
+    // TODO: replace with some robust validation - Joi?
+    if (!req.body.email || !req.body.password) {
+        return res.status(400).send('Email and password must be supplied');
+    }
     const userRepository = getRepository(User);
     const user = new User();
     user.email = req.body.email;
